% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{data}
\name{FREDMD}
\alias{FREDMD}
\title{FRED-MD: A Monthly Database for Macroeconomic Research}
\format{A list with the following:
\describe{
  \item{rawdata:}{The raw dataset as a matrix where the columns are the variables as time series objects.}
  \item{tcodes:}{The transformation codes used to transform the variables to be stationary.}
}}
\source{
\url{http://research.stlouisfed.org/econ/mccracken/fred-md-data.zip}.
}
\usage{
FREDMD
}
\description{
A dataset containing 134 monthly US macroeconomic variables
spanning the period 1959-2014.
}
\references{
McCracken, M. W., Ng, S. (2014) \emph{FRED-MD: A Monthly Database for Macroeconomic Research}. \url{http://www.columbia.edu/~sn2294/papers/freddata.pdf}
}
\keyword{datasets}

