% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lib.R
\name{getmacrodata}
\alias{getmacrodata}
\title{Retrieve transformed dataset for a given date.}
\usage{
getmacrodata(ds, enddate, h = 0, vars = NULL, screen = FALSE)
}
\arguments{
\item{ds}{The raw dataset.}

\item{enddate}{Data until this date is returned.}

\item{h}{Forecast horizon.}

\item{vars}{A vector of names of the variables to be return, or NULL to return all.}

\item{screen}{Should the data be screened for outliers?}
}
\value{
The transformed dataset.
}
\description{
Retrieve transformed dataset for a given date.
}

