% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{data}
\name{JLN2013}
\alias{JLN2013}
\title{US macroeconomic data set from Jurado, Ludvigson and Ng (2013)}
\format{A list with the following:
\describe{
  \item{rawdata:}{The raw dataset as a matrix where the columns are the variables as time series objects.}
  \item{tcodes:}{The transformation codes used to transform the variables to be stationary.}
  \item{names:}{Short names of the variables.}
  \item{desc:}{Description of the variables.}
}}
\source{
\url{http://www.econ.nyu.edu/user/ludvigsons/jlndata.zip}.
}
\usage{
JLN2013
}
\description{
A dataset containing 132 monthly US macroeconomic variables
spanning the period 1960-2011. The dataset is from Jurado, Ludvigson and Ng (2013).
}
\references{
Jurado, K., Ludvigson, S. C., Ng, S. (2013) \emph{Measuring Uncertainty}. \url{http://www.econ.nyu.edu/user/ludvigsons/jln.pdf}
}
\keyword{datasets}

