% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{data}
\name{LN2010}
\alias{LN2010}
\title{US macroeconomic data set from Ludvigson and Ng (2010)}
\format{A list with the following:
\describe{
  \item{rawdata:}{The raw dataset as a matrix where the columns are the variables as time series objects.}
  \item{tcodes:}{The transformation codes used to transform the variables to be stationary.}
}}
\source{
\url{http://www.econ.nyu.edu/user/ludvigsons/Data&ReplicationFiles.zip}.
}
\usage{
LN2010
}
\description{
A dataset containing 131 monthly US macroeconomic variables
spanning the period 1964-2007. The dataset is from Ludvigson and Ng (2010).
}
\references{
Ludvigson, S. C., Ng, S. (2010) \emph{A Factor Analysis of Bond Risk Premia}. Handbook of Empirical Economics and Finance, 2010, e.d. by Aman Uhla and David E. A. Giles, pp. 313-372. Chapman and Hall, Boca Raton, FL. \url{http://www.econ.nyu.edu/user/ludvigsons/handbook-big.pdf}
}
\keyword{datasets}

