% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{package}
\name{macrods-package}
\alias{macrods-package}
\title{macrods: A collection of macroeconomic datasets.}
\description{
Note: The package source contains all the original data sets
as well as the R scripts used to convert them to R format.
}

