% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{data}
\name{SW2002}
\alias{SW2002}
\title{US macroeconomic data set from Stock and Watson (2002)}
\format{A list with the following:
\describe{
  \item{rawdata:}{The raw dataset as a matrix where the columns are the variables as time series objects.}
  \item{tcodes:}{The transformation codes used to transform the variables to be stationary.}
  \item{unbal:}{Boolean vector indicating whether a given variable contains missing values.}
  \item{desc:}{A matrix containing descriptions of the variables.}
}}
\source{
\url{http://www.princeton.edu/~mwatson/ddisk/wc00.zip}.
}
\usage{
SW2002
}
\description{
A dataset containing 215 monthly US macroeconomic variables
spanning the period 1959-1998. The dataset is from Stock and Watson (2002).
}
\references{
Stock, J., Watson, M. (2002). \emph{Macroeconomic forecasting using diffusion indexes}. Journal of Business and Economic Statistics 20(2), 147--162. \url{http://dx.doi.org/10.1198/073500102317351921}
}
\keyword{datasets}

