% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/macrods-package.R
\docType{data}
\name{SW2009}
\alias{SW2009}
\title{US macroeconomic data set from Stock and Watson (2009)}
\format{A list with the following:
\describe{
  \item{rawdata:}{The raw dataset as a matrix where the columns are the variables as time series objects.}
  \item{tcodes:}{The transformation codes used to transform the variables to be stationary.}
  \item{desc:}{A matrix containing descriptions of the variables.}
  \item{shortnames:}{Short names of the variables.}
  \item{longnames:}{Description of the variables.}
}}
\source{
\url{http://www.princeton.edu/~mwatson/ddisk/hendryfestschrift_replicationfiles_April28_2008.zip}.
}
\usage{
SW2009
}
\description{
A dataset containing 187 monthly US macroeconomic variables
spanning the period 1959-2006. The dataset is from Stock and Watson (2009).
}
\references{
Stock, J. and M. Watson (2009). \emph{Forecasting in dynamic factor models subject to structural instability}. In J. Castle and N. Shephard (Eds.), The Methodology and Practice of Econometrics: A Festschrift in Honour of David F. Hendry, pp. 173-205. Oxford University Press.
}
\keyword{datasets}

